Use default name in traces
